[Feat][flink-k8s-v2] Migrate ENABLE_V2 to streampark-common module. #2882
