+++ streampark-common/src/main/scala/org/apache/streampark/common/conf/K8sFlinkConfig.scala
@@ -18,9 +18,21 @@
 package org.apache.streampark.common.conf
 
 /** Flink kubernetes Configuration for v1 version */
-@deprecated("see: org.apache.streampark.flink.kubernetes.v2.Config")
+
 object K8sFlinkConfig {
 
+  lazy val isV2Enabled: Boolean = InternalConfigHolder.get(ENABLE_V2)
+
+  val ENABLE_V2: InternalOption = InternalOption(
+    key = "streampark.flink-k8s.enable-v2",
+    defaultValue = false,
+    classType = classOf[Boolean],
+    description = "Whether to enable the v2 version(base on flink-kubernetes-operator) of flink kubernetes operation"
+  )
+
+  //  ======= deprecated =======
+
+  @deprecated
   val jobStatusTrackTaskTimeoutSec: InternalOption = InternalOption(
     key = "streampark.flink-k8s.tracking.polling-task-timeout-sec.job-status",
     defaultValue = 120L,
@@ -28,6 +40,7 @@ object K8sFlinkConfig {
     description = "run timeout seconds of single flink-k8s metrics tracking task"
   )
 
+  @deprecated
   val metricTrackTaskTimeoutSec: InternalOption = InternalOption(
     key = "streampark.flink-k8s.tracking.polling-task-timeout-sec.cluster-metric",
     defaultValue = 120L,
@@ -35,6 +48,8 @@ object K8sFlinkConfig {
     description = "run timeout seconds of single flink-k8s job status tracking task"
   )
 
+
+  @deprecated
   val jobStatueTrackTaskIntervalSec: InternalOption = InternalOption(
     key = "streampark.flink-k8s.tracking.polling-interval-sec.job-status",
     defaultValue = 5L,
@@ -42,6 +57,7 @@ object K8sFlinkConfig {
     description = "interval seconds between two single flink-k8s metrics tracking task"
   )
 
+  @deprecated
   val metricTrackTaskIntervalSec: InternalOption = InternalOption(
     key = "streampark.flink-k8s.tracking.polling-interval-sec.cluster-metric",
     defaultValue = 5L,
@@ -49,6 +65,7 @@ object K8sFlinkConfig {
     description = "interval seconds between two single flink-k8s metrics tracking task"
   )
 
+  @deprecated
   val silentStateJobKeepTrackingSec: InternalOption = InternalOption(
     key = "streampark.flink-k8s.tracking.silent-state-keep-sec",
     defaultValue = 60,
@@ -69,6 +86,7 @@ object K8sFlinkConfig {
   )
 
   /** kubernetes default namespace */
+  @deprecated
   val DEFAULT_KUBERNETES_NAMESPACE = "default"
 
 }
